refactor: move atomic output-file helpers from command layer to core/io
