+++ src/core/io/mod.rs
@@ -7,5 +7,9 @@
 //! workflows that recurse or loop in the same shape.
 
 pub(crate) mod copy_tree;
+pub mod output_file;
 
 pub(crate) use copy_tree::{copy_tree, EntryPolicy};
+pub use output_file::{
+    write_output_file, write_output_file_atomically, OutputWriteOptions, TrailingNewline,
+};
